data-producer/requirements.txt
kafka-python
lxml
geopy